requirements.txt
ftfy==6.3.1
h5py==3.7.0
imageio==2.19.3
matplotlib==3.5.2
numpy==1.21.5
Pillow==9.2.0
Pillow==12.1.0
regex==2022.7.9
scipy==1.9.1
tabulate==0.8.10
termcolor==3.3.0
timm==1.0.24
torch==2.8.0
torchvision==0.23.0
tqdm==4.64.1
